Adds optional arg in `capacitated_kmeans()` to specify no of clusters
Introduces the `k_spec` parameter to allow users to explicitly specify the number of clusters to be created. If not provided (or 0), the number of clusters is inferred, and able to be increased.
Allows specification of cluster number

Introduces an optional `k_spec` parameter to the clustering functions, allowing users to define the desired number of clusters.

If `k_spec` is not provided (defaults to 0), the algorithm will continue to dynamically determine the number of clusters based on reef density and `max_cluster_size`.