Includes vessel weightings in objective calcs
Allows for the application of weightings to the distance
calculation for each vessel type (mothership, tenders) within the
simulated annealing optimization process.

This provides accurate definition of the objective function w.r.t. weights. This can be used to prioritize utilization of certain vessel types or
account for differences in their operating costs.