Adds functions for superdiagonal matrix operations
Adds functions to create a square distance matrix from a vector of distances, populating the first superdiagonal, and to extract the values from the first superdiagonal of a square matrix.
These functions are used when only the fesaible vector is returned (to save comp time), and returns a distance matrix with only the superdiagonal populated.
This represents a string of sequential distances from start to finish.