Use initial constant change
I suspect there are issues in the critical path calculation process because it is very common to hit `NaN` and `Inf` values in the critical path calculation.

Plotting the provided waypoints when using the default Line Search method, the provided waypoints do not change much compared to the initial guess, indicating the optimiser is struggling to find a suitable gradient.

Changing the naive penalty score approach resolves the main issue of poor convergence behaviour, but the approach still wastes time searching infeasible directions (I suspect again because of the issue in critical path calculation).

Using an alternate constant change approach helps but obviously does not resolve the underlying issue with critical path estimation.

https://julianlsolvers.github.io/LineSearches.jl/v6.0.2/reference/initialstep.html#LineSearches.InitialConstantChange