Reverses path finding if no path found
Improves pathfinding robustness by attempting to find a path in reverse if the initial forward search fails.
The segments of the reversed path are then flipped to return the path in the correct direction.
NOTE: this is a temporary improvement, which does fix *some* problematic scenarios, but is not expected to fix *all* scenarios.