Improves naive score calc to ensure penalty
Updates the naive score calculation to use the maximum vessel weighting rather than the first (mothership) value in the tups.
This ensures a more consistent and accurate penalty when waypoints are excluded during optimization.
Especially in cases where, for some reason, the mothership weighting is *less than* the tender weighting.